feat(quest,karma): quêtes en ordre libre, karma dialogue et combat
Les quêtes ALL_STEPS valident chaque trigger indépendamment (dialogues,
QUEST_DONE, kills). Karma flottant avec −0,15 par monstre, actions KARMA
en dialogue, conditions d’accès et parseur CSV robuste pour le TEXTE.

+++ Global/GameManager.cs
@@ -132,7 +132,7 @@ public partial class GameManager: Node
             deltaRaw = args[1];
         }
 
-        if (!int.TryParse(deltaRaw, NumberStyles.Integer, CultureInfo.InvariantCulture, out int delta))
+        if (!float.TryParse(deltaRaw, NumberStyles.Float, CultureInfo.InvariantCulture, out float delta))
         {
             GD.PrintErr($"[GameManager] Action KARMA : delta invalide '{deltaRaw}'.");
             return;
